Fix issue with GDB breakpoints
This patch fixes an issue with GDB breakpoints where continuing after a
breakpoint would cause an invalid instruction exception. This was due
to how GDB sets soft breakpoints, and how Unicorn records the state
after emulation is halted.

GDB sets soft breakpoints using the single byte `int3` instruction
(CCh).  After the instruction executes, EIP will point to the next
instruction, which is likely to be in the middle of the instruction we
are trying to break on! In this case, roll EIP back by 1 so that when
GDB resumes execution, EIP is will be in the correct place.

An issue with calling kernel functions is also fixed. This was due to
how `hlt` instructions are used to enter kernel functions. If we are
single stepping code, execution will stop before the `hlt` is executed
but the code which looks to see if we are entering a kernel function
would execute and incorrectly calculate the import number.