Allow Super I/O to be enabled on retail systems
Some modified BIOSes run on retail systems and include the kernel debugger that uses the serial port for communication. Having this as an option is very useful, as specifying a debug kit modifies hardware in a way that might be incompatible with those BIOSes.

The Super I/O hardware is always enabled on debug kits.