WIP: Refactor ATA commands into classes
Implement protocols as base classes on top of a common interface so that command processing and usage is uniform. Commands are much easier to implement on this platform, and most of the complicated state management in ATADevice is eliminated.

Next step is to move the DMA operations out of ATADevice and have ATAChannel delegate them to the current command